fix: Move isMacOs to browser.ts to fix SSR error
The isMacOs() function used the `browser` variable from SvelteKit's
$app/environment but the import was missing after a rebase, causing
a ReferenceError during SSR that broke all e2e tests.

Moved isMacOs() from generic.ts to browser.ts for better separation
of concerns and updated all import references.
