+++ src/components/timeline/TimelineInteractionModeControl.svelte
@@ -2,8 +2,7 @@
   import CursorDefaultIcon from '@nasa-jpl/stellar/icons/cursor_default.svg?component';
   import MoveIcon from '@nasa-jpl/stellar/icons/move.svg?component';
   import { createEventDispatcher, onMount } from 'svelte';
-  import { addPageFocusListener } from '../../utilities/browser';
-  import { isMacOs } from '../../utilities/generic';
+  import { addPageFocusListener, isMacOs } from '../../utilities/browser';
   import { isMetaOrCtrlPressed } from '../../utilities/keyboardEvents';
   import { TimelineInteractionMode } from '../../utilities/timeline';
   import { tooltip } from '../../utilities/tooltip';
