+++ src/utilities/generic.test.ts
@@ -11,17 +11,10 @@ import {
   extractQuotes,
   filterEmpty,
   filterNullish,
-  isMacOs,
   lowercase,
   parseJSONStream,
   unique,
 } from './generic';
-
-const mockNavigator = {
-  platform: 'MacIntel',
-};
-
-vi.stubGlobal('navigator', mockNavigator);
 
 describe('Generic utility function tests', () => {
   afterAll(() => {
@@ -126,34 +119,6 @@ describe('Generic utility function tests', () => {
     test('Should return null when attempting to convert non-stringable values', () => {
       expect(attemptStringConversion(null)).toEqual(null);
       expect(attemptStringConversion(undefined)).toEqual(null);
-    });
-  });
-
-  describe('isMacOs', () => {
-    test('Should return true for Mac browsers', () => {
-      expect(isMacOs()).toEqual(true);
-
-      mockNavigator.platform = 'MacPPC';
-      expect(isMacOs()).toEqual(true);
-
-      mockNavigator.platform = 'Mac68K';
-      expect(isMacOs()).toEqual(true);
-    });
-
-    test('Should return false for Windows browsers', () => {
-      mockNavigator.platform = 'Win32';
-      expect(isMacOs()).toEqual(false);
-
-      mockNavigator.platform = 'Windows';
-      expect(isMacOs()).toEqual(false);
-    });
-
-    test('Should return false for Linux browsers', () => {
-      mockNavigator.platform = 'Linux i686';
-      expect(isMacOs()).toEqual(false);
-
-      mockNavigator.platform = 'Linux x86_64';
-      expect(isMacOs()).toEqual(false);
     });
   });
