+++ src/utilities/generic.ts
@@ -140,19 +140,6 @@ export function isEmpty(value: any): boolean {
 }
 
 /**
- * Returns true if the current browser is running on MacOS
- */
-export function isMacOs(): boolean {
-  if (!browser) {
-    return false;
-  }
-  // userAgentData is not yet in TypeScript's Navigator type (only supported in Chromium browsers)
-  const nav = navigator as Navigator & { userAgentData?: { platform: string } };
-  const platform = nav.userAgentData?.platform ?? navigator.platform;
-  return /mac/i.test(platform);
-}
-
-/**
  * Returns true if the mouse event is a right click
  */
 export function isRightClick(e: MouseEvent) {
